requirements.txt
fastapi==0.97.0
hypercorn==0.14.3
sqlalchemy==1.4.25
sqlacodegen==2.3.0
pymysql==1.0.3
requests==2.31.0
pytest-html==3.2.0
python-dotenv==1.0.0
pydantic==1.10.9
pandas==2.0.2
numpy==1.25.0
cryptography==3.4.7
PyJWT==2.7.0
boto3==1.28.1
pytest-mock==3.11.1